feat: 전시 콘텐츠 관리 및 내부 사진 관리 리팩토링
