+++ src/components/artworks-manage/constants.ts
@@ -1,3 +1,6 @@
+import { CONTENT_DESCRIPTION_MAX_LENGTH } from '@/constants/artworkManage';
+
 export const ROW_HEIGHT = 112; // 카드 높이
 export const ROW_GAP = 12;
 export const STEP = ROW_HEIGHT + ROW_GAP;
+export { CONTENT_DESCRIPTION_MAX_LENGTH };
