+++ src/components/artworks-manage/index.ts
@@ -1,3 +1,7 @@
+export { CardPopover } from './CardPopover';
+export { ContentCard } from './ContentCard';
+export { ContentDeleteConfirmDialog } from './ContentDeleteConfirmDialog';
+export { ContentEditSheet } from './ContentEditSheet';
 export { DeleteConfirmDialog } from './DeleteConfirmDialog';
 export { ManageScreen } from './ManageScreen';
 export { OrderScreen } from './OrderScreen';
