+++ src/constants/artworkManage.ts
@@ -1,0 +1,1 @@
+export const CONTENT_DESCRIPTION_MAX_LENGTH = 255;
