+++ src/pages/RegisterCompletePage.tsx
@@ -1,9 +1,10 @@
 import { Check, Info } from 'lucide-react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
+import { BottomButtonBar } from '@/components/common/BottomButtonBar';
 import { VISIBILITY_LABEL, type VisibilityType } from '@/constants/visibility';
 
-interface CompleteState {
+interface ExhibitionCompleteState {
   title?: string;
   school?: string;
   department?: string;
@@ -11,6 +12,13 @@ interface CompleteState {
   placeName?: string;
   artworkVisibility?: VisibilityType;
   contentVisibility?: VisibilityType;
+}
+
+interface ArtworkCompleteState {
+  artworkName?: string;
+  artistName?: string;
+  registrarName?: string;
+  managerName?: string;
 }
 
 function SummaryRow({ label, value }: { label: string; value: string }) {
@@ -22,35 +30,63 @@ function SummaryRow({ label, value }: { label: string; value: string }) {
   );
 }
 
+// ----------------------------------------------------------------------
+// Common Components
+// ----------------------------------------------------------------------
+
+function CompleteHeader({ title, description }: { title: string; description: React.ReactNode }) {
+  return (
+    <div className="mt-40 flex flex-col items-center gap-2.5">
+      <div className="flex size-20 items-center justify-center rounded-full bg-box200">
+        <div className="flex size-12 items-center justify-center rounded-full bg-dark">
+          <Check className="size-6 text-white" strokeWidth={2.8} />
+        </div>
+      </div>
+      <div className="flex flex-col items-center gap-1">
+        <h1 className="typo-body-xl-bold text-center text-main">{title}</h1>
+        <p className="typo-body-xs-regular text-center text-faint">{description}</p>
+      </div>
+    </div>
+  );
+}
+
+function CompleteInfo({ message }: { message: string }) {
+  return (
+    <div className="mt-auto pt-12 pb-6">
+      <div className="flex items-start gap-2 rounded-2xl bg-card p-3.5">
+        <Info className="mt-0.5 size-4 shrink-0 text-faint" strokeWidth={1} />
+        <p className="typo-body-xs-regular text-faint">{message}</p>
+      </div>
+    </div>
+  );
+}
+
+// ----------------------------------------------------------------------
+// Page Components
+// ----------------------------------------------------------------------
+
 export function ExhibitionRegisterComplete() {
   const navigate = useNavigate();
-  const { state } = useLocation() as { state: CompleteState | null };
+  const { state } = useLocation() as { state: ExhibitionCompleteState | null };
 
   const affiliation = [state?.school, state?.organizer ?? state?.department]
     .filter(Boolean)
     .join(' ');
 
   return (
-    <div className="mx-auto flex h-dvh w-96 flex-col bg-page">
-      <div className="flex min-h-0 flex-1 flex-col overflow-y-auto px-5">
-        {/* 완료 안내 */}
-        <div className="mt-40 flex flex-col items-center gap-2.5">
-          <div className="flex size-20 items-center justify-center rounded-full bg-box200">
-            <div className="flex size-12 items-center justify-center rounded-full bg-dark">
-              <Check className="size-6 text-white" strokeWidth={2.8} />
-            </div>
-          </div>
-          <div className="flex flex-col items-center gap-1">
-            <h1 className="typo-body-xl-bold text-center text-main">전시 등록이 완료되었어요</h1>
-            <p className="typo-body-xs-regular text-center text-faint">
+    <div className="mx-auto flex min-h-dvh w-full max-w-md flex-col bg-page">
+      <main className="flex flex-1 flex-col px-5 overflow-y-auto pb-24">
+        <CompleteHeader
+          title="전시 등록이 완료되었어요"
+          description={
+            <>
               이제 관람자가 전시를 발견할 수 있어요.
               <br />
               설정한 공개 시점에 맞춰 노출돼요.
-            </p>
-          </div>
-        </div>
+            </>
+          }
+        />
 
-        {/* 요약 */}
         <div className="mt-14 flex items-start gap-1.5">
           <div className="flex flex-1 flex-col gap-1.5">
             <SummaryRow label="전시명" value={state?.title ?? '-'} />
@@ -68,25 +104,59 @@ export function ExhibitionRegisterComplete() {
             />
           </div>
         </div>
-      </div>
 
-      {/* 하단 */}
-      <div className="shrink-0 px-5 pb-4">
-        <div className="flex items-start gap-2 rounded-2xl bg-card p-3.5">
-          <Info className="size-3.5 shrink-0 text-faint" strokeWidth={1} />
-          <p className="typo-body-xs-regular text-faint">
-            등록 후에도 전시 관리에서 수정할 수 있어요.
-          </p>
-        </div>
+        <CompleteInfo message="등록 후에도 전시 관리에서 수정할 수 있어요." />
+      </main>
 
+      <BottomButtonBar>
         <button
           type="button"
           onClick={() => navigate('/my', { replace: true })}
-          className="typo-body-sm-bold mt-4 h-11 w-full rounded-xl bg-dark text-white"
+          className="w-full h-11 py-3 bg-dark rounded-xl typo-body-sm-bold text-card inline-flex justify-center items-center gap-1.5"
         >
           완료
         </button>
-      </div>
+      </BottomButtonBar>
+    </div>
+  );
+}
+
+export function ArtworkRegisterComplete() {
+  const navigate = useNavigate();
+  const { state } = useLocation() as { state: ArtworkCompleteState | null };
+
+  return (
+    <div className="mx-auto flex min-h-dvh w-full max-w-md flex-col bg-page">
+      <main className="flex flex-1 flex-col px-5 overflow-y-auto pb-24">
+        <CompleteHeader
+          title="작품등록이 완료되었어요"
+          description="등록된 작품은 전시 작업 페이지에서 확인할 수 있어요"
+        />
+
+        <div className="mt-14 flex items-start gap-1.5">
+          <div className="flex flex-1 flex-col gap-1.5">
+            <SummaryRow label="작품명" value={state?.artworkName ?? '-'} />
+            <SummaryRow label="작가명" value={state?.artistName ?? '-'} />
+          </div>
+          <div className="flex flex-1 flex-col gap-1.5">
+            <SummaryRow label="등록자" value={state?.registrarName ?? '-'} />
+            <SummaryRow label="담당자" value={state?.managerName ?? '-'} />
+          </div>
+        </div>
+
+        {/* Info 컴포넌트가 없어도 하단 버튼을 밀어내기 위한 여백용 요소 */}
+        <div className="mt-auto pt-12 pb-6" />
+      </main>
+
+      <BottomButtonBar>
+        <button
+          type="button"
+          onClick={() => navigate(-1)}
+          className="w-full h-11 py-3 bg-dark rounded-xl typo-body-sm-bold text-card inline-flex justify-center items-center gap-1.5"
+        >
+          완료
+        </button>
+      </BottomButtonBar>
     </div>
   );
 }
