+++ src/types/artworkManage.ts
@@ -10,3 +10,13 @@ export interface Work {
   coAuthorUserIds?: number[];
   thumbnail?: string;
 }
+
+export type Content = {
+  id: number;
+  title: string;
+  description: string;
+  photoCount: number;
+  thumbnail?: string;
+};
+
+export const EMPTY_CONTENT: Content = { id: 0, title: '', description: '', photoCount: 0 };
