+++ src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './artworkManage';
 export * from './exhibition';
 export * from './invitation';
 export * from './mypage';
